mangrove.js: Move exported types from Market.ts to types/index.ts
The Offer type was declared both places, but in slightly different ways.
