+++ packages/cleaning-bot/src/MarketCleaner.ts
@@ -1,5 +1,6 @@
 import { logger } from "./util/logger";
-import { Market, Offer } from "@giry/mangrove-js/dist/nodejs/market";
+import { Market } from "@giry/mangrove-js/dist/nodejs/market";
+import { Offer } from "@giry/mangrove-js/dist/nodejs/types";
 import { MgvToken } from "@giry/mangrove-js/dist/nodejs/mgvtoken";
 import { Provider } from "@ethersproject/providers";
 import { BigNumber, BigNumberish } from "ethers";
