+++ packages/mangrove.js/src/mangrove.ts
@@ -1,6 +1,6 @@
 import { addresses, decimals as loadedDecimals } from "./constants";
 import * as eth from "./eth";
-import { BookOptions, Market } from "./market";
+import { Market } from "./market";
 import {
   Provider,
   Signer,
@@ -8,6 +8,7 @@ import {
   Bigish,
   globalConfig,
   CreateSignerOptions,
+  BookOptions,
 } from "./types";
 import * as typechain from "./types/typechain";
 import { MgvToken } from "./mgvtoken";
