+++ packages/obfiller-bot/src/OfferTaker.ts
@@ -1,6 +1,7 @@
 import { logger } from "./util/logger";
 import { sleep } from "@giry/commonlib-js";
-import { Market, Offer } from "@giry/mangrove-js/dist/nodejs/market";
+import { Market } from "@giry/mangrove-js/dist/nodejs/market";
+import { Offer } from "@giry/mangrove-js/dist/nodejs/types";
 import { MgvToken } from "@giry/mangrove-js/dist/nodejs/mgvtoken";
 import { Provider } from "@ethersproject/providers";
 import { BigNumberish } from "ethers";
